IN-649: Update oauth20-provider package to use customizable ttl for access_token
The ttl for the access token is not a common value but is requested for each client & req.
Also, only a valid and positive ttl is added in the expires_in hash of the access_token. If the ttl is negative (to indicate infinite timeout), the expires_in is not populated in the token hash.

Also, fixes broken test cases.