feature/KB-5508: Adding support to pass the client credentials in body params.
This fix is necessary as the exisiting mobile client passes the clientId and clientSecret in the body params instead of inserting in the header with Basic Auth.

Reference to rfc 6749 related to clientId and clientSecret Basic Auth encoding:
https://tools.ietf.org/html/rfc6749\#section-2.3.1
As per the rfc, support for Basic Auth encoding of the clientId and clientSecret in the header is compulsory.
Support for clientId and clientSecret in the body params is optional.

Since it being an optional part of the standard it was not implemented.

But without this fix, the mobile client login is broken. Also, force updating all users to the updated mobile client is not feasible.
Hence, this fix in needed.