feature/IN-776: Support for multiple logins with same credentials at same time.
To add support for this feature, the mecahnism of saving & then retrieving the tokens (access token / refresh token) based on the user_id and client_id has been removed.
The reason being, if we supported that we can never store more than one access & refresh tokens for the same user id and client id.
Since, the access & refresh tokens are saved with expiring nature, saving more than one is not a risk.
Also, the standard doesnt the need to store only one at any point.